+++ common/api/tests/test_recherche_entreprises.py
@@ -6,11 +6,11 @@ import requests
 from common.api.recherche_entreprises import fetch_geo_data_from_siret
 
 
-@patch("common.utils.siret.is_valid_siret", return_value=True)
+@patch("common.utils.siret.is_valid_length_siret", return_value=True)
 class TestFetchGeoDataFromSiret(unittest.TestCase):
 
     @patch("requests.get")
-    def test_fetch_geo_data_from_siret_success(self, mock_get, mock_is_valid_siret):
+    def test_fetch_geo_data_from_siret_success(self, mock_get, mock_is_valid_length_siret):
         canteen_siret = "12345678901234"
         test_cases = [
             {
@@ -91,14 +91,14 @@ class TestFetchGeoDataFromSiret(unittest.TestCase):
             result = fetch_geo_data_from_siret(canteen_siret, response)
             self.assertEqual(result, tc["expected_outcome"])
 
-    def test_fetch_geo_data_from_siret_invalid_siret(self, mock_is_valid_siret):
+    def test_fetch_geo_data_from_siret_invalid_siret(self, mock_is_valid_length_siret):
         response = {}
         canteen_siret = "invalid_siret"
         result = fetch_geo_data_from_siret(canteen_siret, response)
         self.assertIsNone(result)
 
     @patch("requests.get")
-    def test_fetch_geo_data_from_siret_no_results(self, mock_get, mock_is_valid_siret):
+    def test_fetch_geo_data_from_siret_no_results(self, mock_get, mock_is_valid_length_siret):
         mock_response = Mock()
         mock_response.ok = True
         mock_response.json.return_value = {"total_results": 0}
@@ -110,28 +110,28 @@ class TestFetchGeoDataFromSiret(unittest.TestCase):
         self.assertIsNone(result)
 
     @patch("requests.get", side_effect=requests.exceptions.HTTPError)
-    def test_fetch_geo_data_from_siret_http_error(self, mock_get, mock_is_valid_siret):
+    def test_fetch_geo_data_from_siret_http_error(self, mock_get, mock_is_valid_length_siret):
         response = {}
         canteen_siret = "12345678901234"
         result = fetch_geo_data_from_siret(canteen_siret, response)
         self.assertEqual(result, {"siret": canteen_siret})
 
     @patch("requests.get", side_effect=requests.exceptions.ConnectionError)
-    def test_fetch_geo_data_from_siret_connection_error(self, mock_get, mock_is_valid_siret):
+    def test_fetch_geo_data_from_siret_connection_error(self, mock_get, mock_is_valid_length_siret):
         response = {}
         canteen_siret = "12345678901234"
         result = fetch_geo_data_from_siret(canteen_siret, response)
         self.assertEqual(result, {"siret": canteen_siret})
 
     @patch("requests.get", side_effect=requests.exceptions.Timeout)
-    def test_fetch_geo_data_from_siret_timeout(self, mock_get, mock_is_valid_siret):
+    def test_fetch_geo_data_from_siret_timeout(self, mock_get, mock_is_valid_length_siret):
         response = {}
         canteen_siret = "12345678901234"
         result = fetch_geo_data_from_siret(canteen_siret, response)
         self.assertEqual(result, {"siret": canteen_siret})
 
     @patch("requests.get", side_effect=Exception)
-    def test_fetch_geo_data_from_siret_unexpected_exception(self, mock_get, mock_is_valid_siret):
+    def test_fetch_geo_data_from_siret_unexpected_exception(self, mock_get, mock_is_valid_length_siret):
         response = {}
         canteen_siret = "12345678901234"
         result = fetch_geo_data_from_siret(canteen_siret, response)
