+++ common/utils/siret.py
@@ -1,7 +1,31 @@
+from django.core.exceptions import ValidationError
+
+
 def normalise_siret(siret):
     return siret.replace(" ", "").replace("\xa0", "")
 
 
-def is_valid_siret(siret: str) -> bool:
+def is_valid_length_siret(siret: str) -> bool:
     if len(siret) == 14:
         return True
+
+
+def validate_siret(siret):
+    """
+    Performs length and Luhn validation
+    (https://portal.hardis-group.com/pages/viewpage.action?pageId=120357227)
+    """
+    if siret is None or siret == "":
+        return
+    siret = normalise_siret(siret)
+    if not is_valid_length_siret(siret):
+        raise ValidationError("14 caractères numériques sont attendus")
+    odd_digits = [int(n) for n in siret[-1::-2]]
+    even_digits = [int(n) for n in siret[-2::-2]]
+    checksum = sum(odd_digits)
+    for digit in even_digits:
+        checksum += sum(int(n) for n in str(digit * 2))
+    luhn_checksum_valid = checksum % 10 == 0
+
+    if not luhn_checksum_valid:
+        raise ValidationError("Le numéro SIRET n'est pas valide.")
